move Tools script into tests folder
